Modified ChangeLog. Eric Jeschke to Eric R. Jeschke to make it go with the
Modified ChangeLog.
Eric Jeschke to Eric R. Jeschke to make it go  with the tutorials.
Removed the ND_Filter photo edit tutorial from the list. Changed the synopses for the Photo Edit tutorials from request by the author.
Add the copyright as the author wanted it to be. Also changed e-mail to author. Fixed the tutorial hyperlink.