Modified ChangeLog. Anti-Aliased_Treshold was wrong so I changed it to
Modified ChangeLog.
Anti-Aliased_Treshold was wrong so I changed it to Anti-Aliased_Threshold otherwise the tutorial was not loaded on the page.
Made the "Using GAP" a link to /tutorials/Using_GAP/
Fixes in the HTML tags. Try to fix the last classes for the filters.
Fixed a spelling in the tutorial.
Fixed one more class="filter" in the tutorial.
Small fixes nothing that would make a big differense.
Add another <Image> to the tags of filter options to the tutorial.
Fixed the link to channel mixer tutorial now links to /tutorials/Color2BW/#channelmixer.
Fixed the links to the outside HTML examples. Spelling was wrong in step 5 for the class of the filter.
Changed Image->Channel Ops->Offset to Filter->Map->Make Seamless. Fixed the classes for filtering.
Made the "Advanced Animation" a link to /tutorials/Advanced_Animations/