Modified ChangeLog. Changed all & to &amp; so that it would be valid.
Modified ChangeLog.
Changed all & to &amp; so that it would be valid. Moved the <a name=""></a> inside of <div></div> so that they can be valid.
Removed <ul> and </ul> in the index file. They where not valid. Needed <li></li> to be valid.
Removed alt="" on line 42. Validator check.
Add a missing </li> tag.
Changed the </ul> so that it quits before starting <ol> to make this valid.
Fixed so that the links worked. Fixed the original tutorial link so that it is using &amp;