Modified ChangeLog. Changing the include files and doing some changes in
Modified ChangeLog.
Changing the include files and doing some changes in the html files to try and make most of them valid still some that needs more work. Also fixed so that the Anti-Aliased Threshold tutorial is complete with original link and e-mail to author.
Helvetix helped me to add a validation link so that you only need to use this when validating. So the site using this include file can validate it directly from the link.