Created admin/gimp-web-urls-subst which is gimp-web-urls modified for the
* Created admin/gimp-web-urls-subst which is gimp-web-urls
modified for the new syntax.

* Added new substitution programme, programmatic/tools/subst.py
which handles arbitrary text files.  Two potentially significant
effects of using the new substitution mechanism is the loss of the
ability to insert a comment declaring the output is automatically
generated, and run-time is increased due to context free
substitutions and the flat substitution dictionary.