Modified ChangeLog. Updated with some more home: and tutorial: links. All
Modified ChangeLog.
Updated with some more home: and tutorial: links.
All files also have a copyright of the text and images in the tutorial in the top of the tutorial now.
Add the new htrw files to be used in the tutorials. All of them uses the htrw files now.
Removed all the html files in the tutorials.