Modified ChangeLog. Add that the tutorials are done. Fixed a few missing
Modified ChangeLog.
Add that the tutorials are done.
Fixed a few missing tags. Fixed a few missing links. Fixed a few typos.
Removed a few <br /> that was breaking the look of the <pre>.
Changed the code a bit. Fixed a few small problems with the look of the code examples.
Fixed a link to gimp-savvy.
Updated a filter path. Add the perlotine link.
A lot of bug fixes. Add more tags to make it look more like the original tutorial. Fixed a few missing links. Fixed position of paragraphs and add some more paragraphs.
Center the text in the "How Do You Make a New Brush?" step.
Fixed a few missing tags.
Add a few files to the repository that was missing in tutorials.