Modified ChangeLog. Add Mat as the macintosh info maintainer. Fixed broken
Modified ChangeLog.
Add Mat as the macintosh info maintainer.
Fixed broken link to The GIMP v0.54
Removed things that did not have content.
Changed the descriptions of the OS2 screenshots. Add two screenshots from the MacOS stuff.
Changed the bugzilla link to be /bugs/howtos/bugzilla.htrw. Add Mat Caughron to the list of people.
"Bugs" now using /bugs/
"Bugs" now using /bugs/. Changed "GIMP for Macintosh" to be linked to /macintosh/.
Changed the screenshots description.
Removed the link to Bugzilla. It is not Unix specific and is at the moment included in /bugs/.
Is now a section with links and other useful things. This also contains the howtos for bug reporting etc.
Add this from the /bugzilla.htrw source.
Template add.
Wrote this as a quick guide.
Add theese for the bugzilla howtos.
With the help from Mat we got 2 screenshots that we can use.
Removed files.