Modified ChangeLog. Looked through this file and formated the text a bit.
Modified ChangeLog.
Looked through this file and formated the text a bit. Change <A HREF=""></A> to <a href=""></a> that is used for XHTML.
Changed color values for em.caution, em.warn and em.hellbrokeloose to make them more clear on website. Also add bold to make them even more clear. This closes bug: #119479