Modified ChangeLog. Removed the headers for the links in the bottom of the
Modified ChangeLog.
Removed the headers for the links in the bottom of the page and add them to the includes instead.
Fixed a typo.
Add developer.gimp.org to the list of important developer links.
Changed the header a little so it uses Links instead of URLs.
Added the menus to the files to make them work like the others.
Now using the new include files.
Added files wich are going to be used for the source howtos for now.