Add optional "platform" parameter to renderDownload function.
If this parameter is undefined, the function's behaviour is not changed: the
include file is determined by the autodetected platform of the visitor.
If the platform parameter is set, then the include file <platform>.xhtml is used.
If the platform parameter is set to "all", then all platform-specific includes
are used.

A link to show all platforms is shown if platform is undefined or set to any
value except "all".