Added mirror setup hints & removed webmaster@ address
Some of the recently proposed ftp mirrors are lacking README files
and are configured incorrectly. THe hints can help to avoid this.

The webmaster@ alias is most likely not read by anyone anymore,
posting to the gimp-developer list is much more useful for anyone
who wants to contact us. If gimp-web will become active ever again,
we can change it to that list.